Allow customizing the list of monitored network interfaces (#18)
* Allow customizing the list of monitored network interfaces

This is a simple way to customize the selection of network interfaces. If INTERFACE_NAMES_OVERRIDE is set in muninlite.conf, that list is used instead of auto-detection. As muninlite.conf
itself is also a script, it is even possible to write a custom command.
This feature is helpful in environments with virtual machines or containers, like docker or lxc where there are lots of br-*, veth*, lxcbr* etc interfaces where monitoring doesn't make much
sense. I didn't find a way to reliably filter physical interfaces.

* Add examples for config with INTERFACE_NAMES_OVERRIDE

* Fix "parameter not set" error

./muninlite: 651: INTERFACE_NAMES_OVERRIDE: parameter not set

* Make example more portable

Co-authored-by: Kenyon Ralph <kenyon@kenyonralph.com>

* Remove example code as requested

---------

Co-authored-by: Kenyon Ralph <kenyon@kenyonralph.com>